parametrize-tests/pyproject.toml
[project]
name = "parametrize-tests"
version = "0.1.0"
description = "YAML-based test parametrization utilities"
readme = "README.md"
authors = [
    { name = "Noam Shabtai", email = "shabtai.noam@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "deepmerge>=2.0",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
